requirements.txt
streamlit>=1.28.0
openai
python-dotenv>=1.0.0
requests>=2.28.0